backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
google-generativeai
chromadb
langchain
langchain-community
pymupdf
python-docx
python-dotenv
pydantic-settings
pydantic
sse-starlette
